requirements.txt
streamlit
docling
chromadb
python-dotenv
pymupdf
pillow
requests
pdf2image
huggingface_hub
torch
torchvision
sentence-transformers

